Rename folder lib/app/{custom_widgets->widgets}/
