+++ lib/app/pages/settings_page.dart
@@ -12,7 +12,7 @@ import 'package:path_provider/path_provider.dart';
 import 'package:share_plus/share_plus.dart';
 
 import '../cubit/cubits.dart';
-import '../custom_widgets/custom_widgets.dart';
+import '../widgets/widgets.dart';
 import '../models/models.dart';
 import '../services/services.dart';
 import '../utils/utils.dart';
