+++ lib/app/utils/dialogs.dart
@@ -5,7 +5,7 @@ import 'package:ouisync_app/generated/l10n.dart';
 import 'package:ouisync_plugin/ouisync_plugin.dart';
 
 import '../bloc/blocs.dart';
-import '../custom_widgets/custom_widgets.dart';
+import '../widgets/widgets.dart';
 import '../models/models.dart';
 import 'utils.dart';
