+++ lib/app/widgets/bars/search_bar.dart
