+++ lib/app/widgets/dialogs/modal_folder_detail_bottom_sheet.dart
@@ -6,7 +6,7 @@ import 'package:ouisync_plugin/ouisync_plugin.dart';
 import '../../bloc/blocs.dart';
 import '../../pages/pages.dart';
 import '../../utils/utils.dart';
-import '../custom_widgets.dart';
+import '../widgets.dart';
 
 class FolderDetail extends StatefulWidget {
   const FolderDetail({
