+++ lib/app/widgets/items/folder_description.dart
